requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiofiles==23.2.1
    # via aiogram
aiogram==3.10.0
aiohttp==3.9.5
    # via aiogram
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
apscheduler==3.10.4
async-timeout==4.0.3
    # via asyncpg
asyncpg==0.29.0
attrs==23.2.0
    # via aiohttp
certifi==2024.7.4
    # via aiogram
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.7
    # via yarl
magic-filter==1.0.12
    # via aiogram
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pydantic==2.8.2
    # via aiogram
pydantic-core==2.20.1
    # via pydantic
python-dotenv==1.0.1
pytz==2024.1
    # via apscheduler
redis==5.0.7
six==1.16.0
    # via apscheduler
typing-extensions==4.12.2
    # via
    #   aiogram
    #   pydantic
    #   pydantic-core
tzlocal==5.2
    # via apscheduler
yarl==1.9.4
    # via aiohttp
